Update patch set 4
Patch Set 4: Code-Review-1

Obsolete is not the anwser, as tempest and tempest-lib should be co-installable. This is the normal case for code which is made of tempest tests (and needs tempest for discoverability) but needs to be ported away from tempest-lib.

There is the problem with the conflicting binaries. I raised the issue upstream and (on IRC) acknowledged, pending resolution. It's tracked here:
https://bugs.launchpad.net/tempest/+bug/1555825

From the RDO point of view, I personally think that the solution is: temporarily don't install the programs from tempest-lib (after updating tempest-lib to 1.0.0), and have tempest-lib depend on tempest. The upstream solution, if different and if ready before mitaka, can be implemented easily.

Patch-set: 4
Label: Code-Review=-1